+++ CHANGELOG.rst
@@ -3,6 +3,23 @@ Change Log
 ==========
 
 
+0.9
+===
+
+Added
+-----
+
+* Implementation of revised damping parameters as desribed in
+  D. G. A. Smith, L. A. Burns, K. Patkowski, and C. D. Sherrill
+  J. Phys. Chem. Lett., 2016, 7, pp 2197–2203.
+  (Functionality should correspond to V3.2 Rev 0 of original dftd3 code.)
+
+Fixed
+-----
+
+* Routine dftd3_pbc_dispersion delivers now correct stress tensor.
+
+
 0.1
 ===